# Conflicts:
#	features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/fragments/SomListFragment.kt
#	features/transaction/sellerorder/src/test/java/com/tokopedia/sellerorder/list/presentation/viewmodels/SomListViewModelTest.kt
#	features/user/sessioncommon/src/main/java/com/tokopedia/sessioncommon/domain/usecase/GetAdminTypeUseCase.kt

+++ features/top_ads/top_ads_dashboard/src/main/java/com/tokopedia/topads/dashboard/view/adapter/insight/TopadsDailyBudgetRecomAdapter.kt
@@ -61,10 +61,10 @@ class TopadsDailyBudgetRecomAdapter(private val onBudgetClicked: ((pos: Int) -> 
                     setCurrentBid = number
                     holder.view.potentialClick.text = String.format(holder.view.context.getString(R.string.topads_dash_potential_click_text), calculatePotentialClick(holder).thousandFormatted())
                     when {
-                        number < suggestedPriceDaily.toDouble() && number > priceDaily -> {
+                        number < suggestedPriceDaily && number > priceDaily -> {
                             holder.view.buttonSubmitEdit.isEnabled = true
                             holder.view.editBudget?.setError(false)
-                            holder.view.editBudget?.setMessage(String.format(holder.view.context.getString(R.string.topads_dash_budget_recom_error), suggestedPriceDaily))
+                            holder.view.editBudget?.setMessage(String.format(holder.view.context.getString(R.string.topads_dash_budget_recom_error), suggestedPriceDaily.toInt()))
                         }
                         number < priceDaily -> {
                             holder.view.editBudget?.setError(true)
